refactor: Simplified the use of convertArgsToSignature
